application.cpp:  Fix compiler warning

+++ libraries/app/application.cpp
@@ -74,7 +74,7 @@ namespace detail {
       initial_state.initial_timestamp = time_point_sec(time_point::now().sec_since_epoch() /
             initial_state.initial_parameters.block_interval *
             initial_state.initial_parameters.block_interval);
-      for( int i = 0; i < initial_state.initial_active_witnesses; ++i )
+      for( uint64_t i = 0; i < initial_state.initial_active_witnesses; ++i )
       {
          auto name = "init"+fc::to_string(i);
          initial_state.initial_accounts.emplace_back(name,
